test(engine): expand test suites and add new integration tests
New test files (8):
- claim_atom_lattice_integration.rs
- claim_envelope_contract_integration.rs
- hindsight_boundary_capture_integration.rs
- module_live_binding_integration.rs
- persistent_cache_contract.rs
- seqlock_fastpath_integration.rs
- seqlock_rollout_guard.rs
- stage_envelope_certificate_integration.rs

Heavily expanded existing test files with hundreds of new test cases:
- law_mining_cli (+729 lines, new CLI contract tests)
- iterator_protocol_integration (+445 lines)
- seqlock_reader_writer_contract (+460 lines)
- rgc_react_capability_contract (+440 lines)
- rgc_claim_envelope_contract (+397 lines)
- seqlock_candidate_inventory (+391 lines)
- tier_up_profiler_integration (+383 lines, new profiler tests)
- rgc_hindsight_boundary_capture (+375 lines)
- law_mining_integration (+374 lines)
- phase_a_exit_gate_contract (+324 lines)
- docs_help_surface_audit (+269 lines)
- rgc_v8_supremacy_claim_contract (+252 lines)
- Plus clippy lint fixes (vec![] -> []) across all modified test files

Total: ~5,750 new lines of test coverage across 72 test files.

+++ crates/franken-engine/tests/replacement_lineage_log_integration.rs
@@ -199,7 +199,7 @@ fn log_append_multiple() {
         let receipt = make_receipt(
             &format!("old-{i}"),
             &format!("new-{i}"),
-            (i + 1) as u64 * 1_000_000,
+            (i + 1) * 1_000_000,
         );
         let seq = log
             .append(
@@ -224,12 +224,12 @@ fn log_entries_have_hash_chain() {
         let receipt = make_receipt(
             &format!("old-{i}"),
             &format!("new-{i}"),
-            (i + 1) as u64 * 1_000_000,
+            (i + 1) * 1_000_000,
         );
         log.append(
             receipt,
             ReplacementKind::DelegateToNative,
-            (i + 1) as u64 * 1_000_000,
+            (i + 1) * 1_000_000,
         )
         .unwrap();
     }
@@ -269,12 +269,12 @@ fn inclusion_proof_verifies() {
         let receipt = make_receipt(
             &format!("old-{i}"),
             &format!("new-{i}"),
-            (i + 1) as u64 * 1_000_000,
+            (i + 1) * 1_000_000,
         );
         log.append(
             receipt,
             ReplacementKind::DelegateToNative,
-            (i + 1) as u64 * 1_000_000,
+            (i + 1) * 1_000_000,
         )
         .unwrap();
     }
@@ -334,12 +334,12 @@ fn consistency_proof_between_checkpoints() {
         let receipt = make_receipt(
             &format!("old-{i}"),
             &format!("new-{i}"),
-            (i + 1) as u64 * 1_000_000,
+            (i + 1) * 1_000_000,
         );
         log.append(
             receipt,
             ReplacementKind::DelegateToNative,
-            (i + 1) as u64 * 1_000_000,
+            (i + 1) * 1_000_000,
         )
         .unwrap();
     }
@@ -349,12 +349,12 @@ fn consistency_proof_between_checkpoints() {
         let receipt = make_receipt(
             &format!("old-{i}"),
             &format!("new-{i}"),
-            (i + 1) as u64 * 1_000_000,
+            (i + 1) * 1_000_000,
         );
         log.append(
             receipt,
             ReplacementKind::DelegateToNative,
-            (i + 1) as u64 * 1_000_000,
+            (i + 1) * 1_000_000,
         )
         .unwrap();
     }
@@ -375,12 +375,12 @@ fn query_all() {
         let receipt = make_receipt(
             &format!("old-{i}"),
             &format!("new-{i}"),
-            (i + 1) as u64 * 1_000_000,
+            (i + 1) * 1_000_000,
         );
         log.append(
             receipt,
             ReplacementKind::DelegateToNative,
-            (i + 1) as u64 * 1_000_000,
+            (i + 1) * 1_000_000,
         )
         .unwrap();
     }
@@ -488,12 +488,12 @@ fn audit_populated_log() {
         let receipt = make_receipt(
             &format!("old-{i}"),
             &format!("new-{i}"),
-            (i + 1) as u64 * 1_000_000,
+            (i + 1) * 1_000_000,
         );
         log.append(
             receipt,
             ReplacementKind::DelegateToNative,
-            (i + 1) as u64 * 1_000_000,
+            (i + 1) * 1_000_000,
         )
         .unwrap();
     }
@@ -614,12 +614,12 @@ fn full_lifecycle_lineage_log() {
         let receipt = make_receipt(
             &format!("cell-v{i}"),
             &format!("cell-v{}", i + 1),
-            (i + 1) as u64 * 1_000_000,
+            (i + 1) * 1_000_000,
         );
         log.append(
             receipt,
             ReplacementKind::DelegateToNative,
-            (i + 1) as u64 * 1_000_000,
+            (i + 1) * 1_000_000,
         )
         .unwrap();
     }
